requirements.txt
# GCP Cloud Functions 框架
functions-framework==3.*

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Google AI (Gemini)
google-generativeai>=0.8.3

# Anthropic Claude (optional: set LLM_PROVIDER=claude in .env to use)
anthropic>=0.40.0

# Google APIs (Calendar, Sheets)
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread

# 工具類
python-dotenv
pytz
requests
